Working on octtree
