--- lib/src/Shapes/Octtree/Leaf.dart
@@ -1,8 +1,0 @@
-part of ThreeDart.Shapes.Octtree;
-
-class Leaf extends Node {
-
-  Leaf(): super._() {
-    
-  }
-}
