--- lib/src/Shapes/Octtree/Node.dart
@@ -1,7 +1,0 @@
-part of ThreeDart.Shapes.Octtree;
-
-class Node {
-
-  Node._();
-
-}
